prompt2song_service/pyproject.toml
[project]
name = "prompt2song"
version = "0.1.0"
description = "Prompt-to-song API and utilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "pyyaml",
    "torch",
    "transformers",
]

[project.optional-dependencies]
dev = ["ipykernel"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["prompt2song"]
